add themes file folder in daijiang.github.io
I need to put themes file folder in the top folder, which is
daijiang.github.io. Since in the zoey bolg, all CSS were referred to
/themes/violet/css/style.css. On github webpage, it will look for these
css in the folder of the top folder. Thus, I need to put the themes
folder in my daijiang.github.io. Another choice is to change every
/themes/violet to /zoey/themes/violet, which is unpreferable.